Move createDefaultEditingStyles() to ol.style
